requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
langchain==0.3.8
pydantic==2.8.2
langchain-openai==0.2.9
motor==3.3.1
pymongo==4.6.1
